% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3.R
\name{.callP3NreadOrg}
\alias{.callP3NreadOrg}
\title{call primer3 for a given set of DNAstringSet object}
\usage{
.callP3NreadOrg(
  seq,
  size_range = "200-700",
  Tm = c(58, 60, 70),
  name = "Hodor",
  primer3 = NULL,
  thermo.param = NULL,
  settings = NULL,
  excluded_regions = NULL
)
}
\arguments{
\item{seq}{DNAstring object, one DNA string for the given amplicon}

\item{size_range}{default: '151-500'}

\item{Tm}{melting temprature parameters default:c(55,57,58)}

\item{name}{name of the amplicon in chr_start_end format}

\item{primer3}{primer3 location}

\item{settings}{text file for parameters}

\item{therme.param}{thermodynamic parameters folder}
}
\description{
call primer3 for a given set of DNAstringSet object
}
\examples{

primers=mapply(.callP3NreadOrg,seq=sample.seq,
                                  name=names(sample.seq),
                     MoreArgs=list(primer3=primer3,size_range=size_range,Tm=Tm,
                                   thermo.param=thermoParam,settings=settings)
                     ,SIMPLIFY = FALSE)


}
\author{
Altuna Akalin modified Arnaud Krebs' original function
}
