% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{single_primer_design}
\alias{single_primer_design}
\title{Initiate the primer design for a single gene}
\usage{
single_primer_design(
  ncbi = NULL,
  transcript_id_version = NULL,
  primer3 = NULL,
  thermo.param = system.file("extdata/primer3_config/", package = "scAmpDesign"),
  settings = system.file("extdata/primer3_settings.txt", package = "scAmpDesign"),
  ensembl_host = "http://uswest.ensembl.org/",
  dataset = "hsapiens_gene_ensembl"
)
}
\arguments{
\item{ncbi}{An NBCI Reference. For example, the SARS-COV2 is NC_045512.}

\item{transcript_id_version}{An Ensembl transcript id with version number. For example,
the main transcript for FOXP3 is ENST00000376207.10.}

\item{primer3}{The location of primer3_core. Example:  primer3='/Users/adamcribbs/miniconda3/bin/primer3_core'}

\item{thermo.param}{The location of the thermodynamic parameters folder. Currently within the R inst/extdata}

\item{settings}{text file for parameters for primer 3. Currently within the inst/extdata}

\item{ensembl_host}{Specify the host for ensembl. Defaults to "http://uswest.ensembl.org/".}

\item{ensembl_dataset}{Specify the datasdets to be accessed by ensembl. Defaults to "hsapiens_gene_ensembl"}
}
\value{
A set of primer pairs
}
\description{
This function will design a primer set to use with targeted single-cell sequencing.
}
\examples{
single_primer_design(ncbi="NC_045512", primer3="/Users/adamcribbs/miniconda3/bin/primer3_core")
single_primer_design(transcript_id_version="ENST00000376207.10", primer3="/Users/adamcribbs/miniconda3/bin/primer3_core")
}
